requirements.txt
# Important
Django~=5.0.0
django-environ==0.11.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==22.0.0
psycopg2==2.9.9


# Development (You can uncomment these for development)
# pre-commit==3.7.1